fix fwo validation tin

+++ BearGoodbyeKolkhoz.API/Models/OutputModels/CompanyOutputModel.cs
@@ -4,7 +4,7 @@
     {   
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Tin { get; set; }
+        public long Tin { get; set; }
         public string Email { get; set; }
         public string PhoneNumber { get; set; }
         public bool IsDeleted { get; set; }
